requirements/base.txt
Django==2.2.2
psycopg2==2.7.5
django-environ==0.4.5
whitenoise==4.1.2
brotlipy==0.7.0
django-model-utils==3.1.1
djangorestframework==3.9.4
djangorestframework-datatables
django-filter==1.1.0
drf_ujson==1.2.0
unicode-slugify==0.1.3
django-autoslug==1.9.4
pytz==2019.1
kombu==4.5.0
msgpack_python==0.5.6
django-extensions==2.1.7
django-watchman==0.16.0
django-debug-toolbar==1.11
mysqlclient==1.4.6
django-admin-list-filter-dropdown==1.0.2
django-comments-xtd==2.4.0
django-contrib-comments==1.9.1
django-simple-history==2.7.2
django-computedfields==0.0.17
django-crispy-forms==1.7.2
django-crispy-bulma==0.2
django-tinymce==3.0.2
